+++ objectified-ui/public/WHATS_NEW.md
@@ -8,6 +8,11 @@ We continue to improve the platform based on your feedback with improvements and
 
 - Bug fixes:
   - Improves drag/drop behavior in class nodes.
+  - Fixed property template list to show [optional] when nullable is selected.
+  - Fixed property to show numeric precision when applicable.
+- UI Improvements:
+  - Class Template Library now shows the schema preview.
+  - Properties Template Form now includes a "close" button, and allows multiple properties to be added without closing the form.
 
 View our YouTube channel [here](https://www.youtube.com/@objectifieddev) for detailed tutorials and walkthroughs!